Finish bounding box implementation
